More bug fixes and improvements

+++ utilities/util_manga.py
@@ -11,69 +11,87 @@ from utilities.util_network import better_get
 
 
 
-def save_config(key:str=None, value:str=None, replace_data:dict=None, reread_cache:bool=True):
+def save_config(key:str=None, value:str=None, replace_data:bool=False):
     config_path = "./cache/reading_library.json"    
-    if replace_data == None:
-        data = st.session_state.manga_cache
-        if key in data:
-            if isinstance(data[key], dict) and isinstance(value, dict):
-                data[key].update(value) 
+    if not replace_data and key != None and value != None:
+        if key in st.session_state.manga_cache:
+            if isinstance(st.session_state.manga_cache[key], dict) and isinstance(value, dict):
+                st.session_state.manga_cache[key].update(value) 
             else:
-                data[key] = value
+                st.session_state.manga_cache[key] = value
         else:
-            data[key] = value
+            st.session_state.manga_cache[key] = value
 
         with open(config_path, "w") as f:
-            json.dump(data, f, indent=4)
+            json.dump(st.session_state.manga_cache, f, indent=4)
     
     else:
         with open(config_path, "w") as f:
-            json.dump(replace_data, f, indent=4)
-    
-    if reread_cache:
-        st.session_state.manga_cache = read_cache()
+            json.dump(st.session_state.manga_cache, f, indent=4)
 
 
 
 
-def refresh_library(data:list, title:str=None) -> None:
-    if title == None:
-        for key, value in data.items():
-            if data[key]["website"] == "asurascans.com/":
-                refresh_asura(key, value)
-            elif data[key]["website"] == "mangadex.org/":
+def refresh_library(title:str=None) -> None:
+    if title is None:
+        for chapter_title, value in st.session_state.manga_cache.items():
+            if value["website"] == "asurascans.com/":
+                chapter_json = asura_get_chapter(chapter_url=value["main_url"], website=value["website"]
+                )
+            elif value["website"] == "mangadex.org/":
                 ...
-            
+            if chapter_json is None:
+                st.toast(body=f":red[Failed to refresh library for {chapter_title}]", duration="infinite", icon=":material/apps_outage:")
+            else:
+                save_config(key=chapter_title, value=chapter_json)
+                st.toast(body=f":blue[Refreshed library for **{chapter_title}**!]", duration="short", icon=":material/cached:")
     else:
-        if data[title]["website"] == "asurascans.com/":
-            refresh_asura(title, data[title])
-        elif data[title]["website"] == "mangadex.org/":
+        if st.session_state.manga_cache[title]["website"] == "asurascans.com/":
+            chapter_json = asura_get_chapter(
+                chapter_url=st.session_state.manga_cache[title]["main_url"],
+                website=st.session_state.manga_cache[title]["website"]
+            )
+        elif st.session_state.manga_cache[title]["website"] == "mangadex.org/":
             ...
+        if chapter_json is None:
+            st.toast(body=f":red[Failed to refresh library for {title}]", duration="infinite", icon=":material/apps_outage:")
+        else:
+            save_config(key=title, value=chapter_json)
+            st.toast(body=f":blue[Refreshed library for **{title}**!]", duration="short", icon=":material/cached:")
 
 
 
-def refresh_asura(key:str, value:dict) -> None:
-    title_url_response = BeautifulSoup(better_get(value["main_url"]).text, "html.parser")
+
+def asura_get_chapter(chapter_url:str, website:str) -> dict:
+    response = better_get(chapter_url)
+    if response is None:
+        return None
     
+    title_url_response = BeautifulSoup(response.text, "html.parser")
     title_chapters_unclean = title_url_response.find("div", class_="divide-y divide-white/5").find_all("a", href=True)
     title_chapters_cleaned = []
     for chapter in title_chapters_unclean[::-1]:
-        title_chapters_cleaned.append(urljoin(f"https://{value['website']}", chapter['href']))
+        title_chapters_cleaned.append(urljoin(f"https://{website}", chapter['href']))
         
     title_status = title_url_response.find("span", class_="text-base font-bold text-[#A78BFA] capitalize").text.strip().capitalize()
+    title_type = title_url_response.find("span", class_="text-base font-bold text-[#913FE2] uppercase").text.strip().capitalize()
     title_rating = title_url_response.find("span", class_="text-xl font-bold bg-gradient-to-r from-[#FFDA6E] to-[#FFC414] bg-clip-text text-transparent").text.strip()
     title_chapter = title_url_response.find("span", class_="text-xl font-bold bg-gradient-to-b from-[#48C855] to-[#C6FFAB] bg-clip-text text-transparent").text.strip()
     title_image = title_url_response.find("img", id="cover-viewer-img")["data-full-src"]
 
-    json_to_save = {
+    chapter_json = {
+        "main_url": chapter_url,
         "chapters_amount": int(title_chapter),
         "status": title_status,
+        "type": title_type,
         "rating": float(title_rating),
+        "website": website,
         "image": title_image,
+        "chapter_downloaded": [],
         "chapters_url": title_chapters_cleaned,
     }
-    save_config(key, json_to_save)
-    st.toast(body=f":blue[Refreshed library for **{key}**!]", duration="short", icon=":material/cached:")
+    return chapter_json
+
 
 
 
@@ -89,33 +107,41 @@ def search_titles(websites:list, title:str) -> list:
             search_response = search_titles_asura(title)
         elif website == "😺 MangaDex":
             search_response = search_titles_mangadex(title)
-        search_results.extend(search_response.keys()) if search_response != {} else None
-        combined_results.update(search_response)
+        if search_response is not None:
+            search_results.extend(search_response.keys())
+            combined_results.update(search_response)
     
-    st.session_state.search_lookup.update(combined_results)
+    if search_results == [] or combined_results == {}:
+        st.session_state.search_lookup = {}
+        return []
+    
+    st.session_state.search_lookup = combined_results
     return search_results
 
 
 
 
 def search_titles_asura(title:str) -> dict:
-    try:
-        asura_search_url = f"https://api.asurascans.com/api/search?q={title}"
-        search_result_unclean = better_get(asura_search_url).json()["data"]
-        search_result_clean = {f"🌑 {item['title']}" : f"https://asurascans.com/comics/{item['slug']}" for item in search_result_unclean}
-        if len(search_result_clean) == 0:
-            raise Exception()
-        else:
-            return search_result_clean
-    except:
-        print("[Error] Result are none, please try with another title")
-        return {}
+    response = better_get(f"https://api.asurascans.com/api/search?q={title}")
+    if response is None:
+        return None
+    
+    response_json = response.json()["data"]
+    if response_json is None:
+        return None
+    
+    st.write(response_json)
+    search_result_clean = {f"🌑 {item['title']}" : f"https://asurascans.com/comics/{item['slug']}" for item in response_json}
+    if len(search_result_clean) == 0:
+        return None
+    else:
+        return search_result_clean
 
 
 
 
 def search_titles_mangadex(title:str) -> dict:
-    ...
+    return None
 
 
 
@@ -159,11 +185,16 @@ def download_chapter(title:str, chapter_key:str, chapter_url:str, website_type:s
         try:
             images_combine.append(Image.open(image))
         except Exception as e:
-            print(f"[Error] {e} {image}")
             st.toast(body=f":red[Failed to download **Chapter {chapter_key}**]", duration="infinite", icon=":material/file_download_off:")
+            shutil.rmtree(chapter_path, ignore_errors=True)
             return None
     
     pdf_path = f"{title_path}/Chapter {chapter_key.zfill(2)}.pdf"
+    if not images_combine:
+        st.toast(body=f":red[No images found for **Chapter {chapter_key}**]", icon=":material/no_photography:")
+        shutil.rmtree(chapter_path, ignore_errors=True)
+        return None
+    
     images_combine[0].save(pdf_path, "PDF", resolution=100, save_all=True, append_images=images_combine[1:])
     st.toast(body=f":green[Done downloading **Chapter {chapter_key}**]", duration="short", icon=":material/download_done:")
     
@@ -172,13 +203,16 @@ def download_chapter(title:str, chapter_key:str, chapter_url:str, website_type:s
     except OSError as e:
         print(f"[Erorr] Failed to remove folder {title}/{chapter_key} - {e}")
     st.session_state.manga_cache[title]["chapter_downloaded"].append(chapter_url)
-    save_config(title, st.session_state.manga_cache[title], reread_cache=False)
+    save_config(replace_data=True)
 
 
 
 
 def get_images_asura(chapter_url:str, chapter_path:str) -> tuple[list[str], list[str]]:
-    chapter_html = BeautifulSoup(better_get(chapter_url).text, "html.parser")
+    response = better_get(chapter_url)
+    if response is None:
+        return None, None
+    chapter_html = BeautifulSoup(response.text, "html.parser")
     chapter_body = chapter_html.find("div", class_="max-w-full md:max-w-[720px] mx-auto overflow-hidden flex flex-col leading-[0]")
     
     chapter_images = chapter_body.find_all("img")    
@@ -196,7 +230,7 @@ def get_images_asura(chapter_url:str, chapter_path:str) -> tuple[list[str], list
 
 def threaded_download(url:str, image_path:str) -> None:
     image_data = better_get(url)
-    if image_data == None:
+    if image_data is None:
         print(url)
     else:
         with open(image_path, "wb") as handler:
@@ -207,22 +241,8 @@ def threaded_download(url:str, image_path:str) -> None:
 
 def change_chapter_read(title:str, chapter_read:int) -> None:
     st.session_state.manga_cache[title]["chapter_read"] = chapter_read
-    save_config(title, st.session_state.manga_cache[title], reread_cache=False)
+    save_config(replace_data=True)
     st.toast(body=":green[Done saving **Chapter Read**]", duration="short", icon=":material/save_as:")
-
-
-
-
-def change_chapter_state(current_chapter:str):
-    st.session_state.open_chapter = current_chapter
-
-
-
-
-@st.cache_data(persist="disk", show_spinner=True)
-def get_cached_image(url:str):
-    response = better_get(url)
-    return response.content
 
 
 
@@ -230,5 +250,5 @@ def get_cached_image(url:str):
 def sync_and_save(new_layout):
     sorted_layout = sorted(new_layout, key=lambda item: (item['y'], item['x']))
     new_order_indices = [int(item['i']) for item in sorted_layout]
-    ordered_data = {list(st.session_state.temp_manga_cache.items())[i][0]: list(st.session_state.temp_manga_cache.items())[i][1] for i in new_order_indices}
-    save_config(replace_data=ordered_data)
+    st.session_state.manga_cache = {list(st.session_state.temp_manga_cache.items())[i][0]: list(st.session_state.temp_manga_cache.items())[i][1] for i in new_order_indices}
+    save_config(replace_data=True)
